update security model

+++ src/main/java/fr/istic/web/rest/StudentResource.java
@@ -3,6 +3,7 @@ package fr.istic.web.rest;
 import static javax.ws.rs.core.UriBuilder.fromPath;
 
 import fr.istic.service.StudentService;
+import fr.istic.web.rest.errors.AccountResourceException;
 import fr.istic.web.rest.errors.BadRequestAlertException;
 import fr.istic.web.util.HeaderUtil;
 import fr.istic.web.util.ResponseUtil;
@@ -12,8 +13,14 @@ import org.eclipse.microprofile.config.inject.ConfigProperty;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import fr.istic.domain.Authority;
+import fr.istic.domain.Course;
+import fr.istic.domain.ExamSheet;
+import fr.istic.domain.Student;
+import fr.istic.domain.User;
 import fr.istic.security.AuthoritiesConstants;
 import fr.istic.service.Paged;
+import fr.istic.service.SecurityService;
 import fr.istic.web.rest.vm.PageRequestVM;
 import fr.istic.web.rest.vm.SortRequestVM;
 import fr.istic.web.util.PaginationUtil;
@@ -46,6 +53,12 @@ public class StudentResource {
 
     @Inject
     StudentService studentService;
+
+    @Inject
+    SecurityService securityService;
+
+
+
     /**
      * {@code POST  /students} : Create a new student.
      *
@@ -75,11 +88,15 @@ public class StudentResource {
      */
     @PUT
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
-    public Response updateStudent(@Valid StudentDTO studentDTO) {
+    public Response updateStudent(@Valid StudentDTO studentDTO, @Context SecurityContext ctx) {
         log.debug("REST request to update Student : {}", studentDTO);
         if (studentDTO.id == null) {
             throw new BadRequestAlertException("Invalid id", ENTITY_NAME, "idnull");
         }
+        if (!securityService.canAccess(ctx, studentDTO.id, Student.class  )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        };
+
         var result = studentService.persistOrUpdate(studentDTO);
         var response = Response.ok().entity(result);
         HeaderUtil.createEntityUpdateAlert(applicationName, true, ENTITY_NAME, studentDTO.id.toString()).forEach(response::header);
@@ -95,8 +112,12 @@ public class StudentResource {
     @DELETE
     @RolesAllowed({AuthoritiesConstants.USER, AuthoritiesConstants.ADMIN})
     @Path("/{id}")
-    public Response deleteStudent(@PathParam("id") Long id) {
+    public Response deleteStudent(@PathParam("id") Long id, @Context SecurityContext ctx) {
         log.debug("REST request to delete Student : {}", id);
+        if (!securityService.canAccess(ctx, id, Student.class )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        };
+
         studentService.delete(id);
         var response = Response.noContent();
         HeaderUtil.createEntityDeletionAlert(applicationName, true, ENTITY_NAME, id.toString()).forEach(response::header);
@@ -111,9 +132,8 @@ public class StudentResource {
      * @return the {@link Response} with status {@code 200 (OK)} and the list of students in body.
      */
     @GET
-    public Response getAllStudents(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @QueryParam(value = "eagerload") boolean eagerload) {
+    public Response getAllStudents(@BeanParam PageRequestVM pageRequest, @BeanParam SortRequestVM sortRequest, @Context UriInfo uriInfo, @QueryParam(value = "eagerload") boolean eagerload, @Context SecurityContext ctx) {
         log.debug("REST request to get a page of Students");
-
         var page = pageRequest.toPage();
 
         var sort = sortRequest.toSort();
@@ -122,17 +142,33 @@ public class StudentResource {
         MultivaluedMap param =  uriInfo.getQueryParameters();
         if (param.containsKey("courseId")){
             List id = (List) param.get("courseId");
+            if (!securityService.canAccess(ctx, Long.parseLong(""+id.get(0)), Course.class )){
+                return Response.status(403, "Current user cannot access to this ressource").build();
+            };
             result =studentService.findStudentsbyCourseId(page, Long.parseLong("" + id.get(0)));
         } else if (param.containsKey("sheetId")){
             List id = (List) param.get("sheetId");
             result =studentService.findStudentsbySheetId(page, Long.parseLong("" + id.get(0)));
         }
-
         else {
-            if (eagerload) {
-                result = studentService.findAllWithEagerRelationships(page);
+            var userLogin = Optional
+            .ofNullable(ctx.getUserPrincipal().getName());
+        if (!userLogin.isPresent()){
+            throw new AccountResourceException("Current user login not found");
+        }
+        var user = User.findOneByLogin(userLogin.get());
+        if (!user.isPresent()) {
+            throw new AccountResourceException("User could not be found");
+        }
+            else if (user.get().authorities.size() >= 1 && user.get().authorities.stream().anyMatch(e1-> e1.equals(new Authority("ROLE_ADMIN")))){
+                if (eagerload) {
+                    result = studentService.findAllWithEagerRelationships(page);
+                } else {
+                    result = studentService.findAll(page);
+                }
+
             } else {
-                result = studentService.findAll(page);
+                return Response.status(403, "Current user cannot access to this ressource").build();
             }
         }
         var response = Response.ok().entity(result.content);
@@ -150,8 +186,11 @@ public class StudentResource {
     @GET
     @Path("/{id}")
 
-    public Response getStudent(@PathParam("id") Long id) {
+    public Response getStudent(@PathParam("id") Long id, @Context SecurityContext ctx ) {
         log.debug("REST request to get Student : {}", id);
+        if (!securityService.canAccess(ctx, id, Student.class )){
+            return Response.status(403, "Current user cannot access to this ressource").build();
+        };
         Optional<StudentDTO> studentDTO = studentService.findOne(id);
         return ResponseUtil.wrapOrNotFound(studentDTO);
     }
